admin verified & count added on admin side

+++ public/admin/manage-users.php
@@ -10,25 +10,33 @@ if (!isset($_SESSION["admin_loggedin"])) {
 require_once('../../app/db/index.php');
 
 if ($stmt1 =
-    $conn->prepare('SELECT id, name, email, mobile, bloodgroup, district, lastdonated, dob, timestamps FROM users ')
+    $conn->prepare('SELECT id, name, email, mobile, bloodgroup, district, lastdonated, dob, timestamps, admin_verified, donation_count FROM users ')
 ) {
     $stmt1->execute();
     $stmt1->store_result();
-    $stmt1->bind_result($id, $name, $email, $mobile, $bloodgroup, $district, $lastdonated, $dob, $timestamps);
+    $stmt1->bind_result($id, $name, $email, $mobile, $bloodgroup, $district, $lastdonated, $dob, $timestamps, $admin_verified, $donation_count);
 }
 
-if (isset($_GET["id"])) {
-    if ($stmt2 = $conn->prepare("DELETE FROM users WHERE id = ?")) {
-        $stmt2->bind_param("i", $_GET["id"]);
-        $stmt2->execute();
-        if ($stmt2->affected_rows == 1) {
-            $stmt3 = $conn->prepare("DELETE FROM blood_records WHERE userid = ?");
-            $stmt3->bind_param("i", $_GET["id"]);
-            $stmt3->execute();
-            header("Location: ./manage-users.php?status=success&message=Successfully deleted 1 user and their blood records.");
-            return;
+if (isset($_GET["id"]) && isset($_GET["action"])) {
+    if ($_GET["action"] === "delete") {
+        if ($stmt2 = $conn->prepare("DELETE FROM users WHERE id = ?")) {
+            $stmt2->bind_param("i", $_GET["id"]);
+            $stmt2->execute();
+            if ($stmt2->affected_rows == 1) {
+                $stmt3 = $conn->prepare("DELETE FROM blood_records WHERE userid = ?");
+                $stmt3->bind_param("i", $_GET["id"]);
+                $stmt3->execute();
+                header("Location: ./manage-users.php?status=success&message=Successfully deleted 1 user and their blood records.");
+                return;
+            }
+            header("Location: ./manage-users.php?status=failure&message=Couldn't delete that user.");
         }
-        header("Location: ./manage-users.php?status=failure&message=Couldn't delete that user.");
+    } elseif ($_GET["action"] === "verify") {
+        $stmt2 = $conn->prepare("UPDATE users SET admin_verified = ? WHERE id = ?");
+        $val = 1;
+        $stmt2->bind_param("ii", $val, $_GET["id"]);
+        $stmt2->execute();
+        header("Location: ./manage-users.php?status=success&message=Successfully verified user");
     }
 }
 ?>
@@ -67,6 +75,8 @@ if (isset($_GET["id"])) {
                     <th>Last Donated</th>
                     <th>Date of birth</th>
                     <th>User Since</th>
+                    <th>Verified User</th>
+                    <th>Donation Count</th>
                     <th>Actions</th>
                 </tr>
                 <?php
@@ -82,7 +92,13 @@ if (isset($_GET["id"])) {
                                     <td>' . $lastdonated . '</td>
                                     <td>' . $dob . '</td>
                                     <td>' . $timestamps . '</td>
-                                    <td><a href="./manage-users.php?id=' . $id . '">Delete</a></td>
+                                    <td>' . ($admin_verified ? "Yes" : "No") . '</td>
+                                    <td>' . $donation_count . '</td>
+                                    <td style="width: 100px">
+                                        <a href="./manage-users.php?id=' . $id . '&action=delete">Delete</a>
+                                        </br>
+                                        ' . (!$admin_verified ? "<a href=./manage-users.php?id=$id&action=verify>Verify User</a>" : "") . '
+                                    </td>
                                 </tr>';
                     $count++;
                 }
